A different implementation of the python solution
the idea is to match the pattern above, then take the line with the pattern, split it by the period(.),
take the word i.e. 1. Angry, when split, becomes ['1', 'angry'], the subject is the second element of the
split and subjectnumber is the first element of the split. For lines with '=', we split by the '=' and take
the first element as the synonym and the second as the meaning For lines with ':', we split by the ':' and
take the second as the example. It turns out there are lines where the subject is in a new line, so for that
we pick those words that do not fit our other conditions and check if we have a subject in the current list,
if not, we push the word